--- a/pkg/statsd/handler_tags.go
+++ b/pkg/statsd/handler_tags.go
@@ -33,7 +33,7 @@ func (th *TagHandler) DispatchMetric(ctx context.Context, m *gostatsd.Metric) er
 	if m.Hostname == "" {
 		m.Hostname = string(m.SourceIP)
 	}
-	m.Tags = append(m.Tags, th.tags...)
+	m.Tags = uniqueTags(m.Tags, th.tags)
 	return th.metrics.DispatchMetric(ctx, m)
 }
 
@@ -42,11 +42,32 @@ func (th *TagHandler) DispatchEvent(ctx context.Context, e *gostatsd.Event) erro
 	if e.Hostname == "" {
 		e.Hostname = string(e.SourceIP)
 	}
-	e.Tags = append(e.Tags, th.tags...)
+	e.Tags = uniqueTags(e.Tags, th.tags)
 	return th.events.DispatchEvent(ctx, e)
 }
 
 // WaitForEvents waits for all event-dispatching goroutines to finish.
 func (th *TagHandler) WaitForEvents() {
 	th.events.WaitForEvents()
+}
+
+func uniqueTags(t1 gostatsd.Tags, t2 gostatsd.Tags) gostatsd.Tags {
+	tags := gostatsd.Tags{}
+	seen := map[string]bool{}
+
+	for _, v := range t1 {
+		if seen[v] != true {
+			tags = append(tags, v)
+			seen[v] = true
+		}
+	}
+
+	for _, v := range t2 {
+		if seen[v] != true {
+			tags = append(tags, v)
+			seen[v] = true
+		}
+	}
+
+	return tags
 }