--- a/pkg/statsd/handler_tags.go
+++ b/pkg/statsd/handler_tags.go
@@ -13,7 +13,7 @@ type TagHandler struct {
 	estimatedTags int
 }
 
-// NewTagHandler initialises a new handler which adds tags and sends metrics/events to the next handler
+// NewTagHandler initialises a new handler which adds unqiue tags and sends metrics/events to the next handler
 func NewTagHandler(metrics MetricHandler, events EventHandler, tags gostatsd.Tags) *TagHandler {
 	return &TagHandler{
 		metrics:       metrics,
@@ -28,7 +28,7 @@ func (th *TagHandler) EstimatedTags() int {
 	return th.estimatedTags
 }
 
-// DispatchMetric adds the tags from the TagHandler to the metric and passes it to the next stage in the pipeline
+// DispatchMetric adds the unqiue tags from the TagHandler to the metric and passes it to the next stage in the pipeline
 func (th *TagHandler) DispatchMetric(ctx context.Context, m *gostatsd.Metric) error {
 	if m.Hostname == "" {
 		m.Hostname = string(m.SourceIP)
@@ -37,7 +37,7 @@ func (th *TagHandler) DispatchMetric(ctx context.Context, m *gostatsd.Metric) er
 	return th.metrics.DispatchMetric(ctx, m)
 }
 
-// DispatchEvent adds the tags from the TagHandler to the event and passes it to the next stage in the pipeline
+// DispatchEvent adds the unqiue tags from the TagHandler to the event and passes it to the next stage in the pipeline
 func (th *TagHandler) DispatchEvent(ctx context.Context, e *gostatsd.Event) error {
 	if e.Hostname == "" {
 		e.Hostname = string(e.SourceIP)