We sometimes receive duplicate tags for a metric. This PR will only add unique tags to the metric or event.

The other approach would be to consolidate the dupe tags into one tag. e.g. from `my_tag:my_value,my_tag:my_value,my_tag:my_value` to `my_tag:my_value__my_value__my_value`. Thoughts on this?