@juliawong 2018-08-03
# Benchmark results

1. `go test -v -run=XXX -bench=BenchmarkTagMetricHandlerAddsDuplicateTags* -count=1 ./pkg/statsd`

Before:
```
goarch: amd64
pkg: github.com/atlassian/gostatsd/pkg/statsd
BenchmarkTagMetricHandlerAddsDuplicateTagsSmall-8   	10000000	       163 ns/op	     170 B/op	       1 allocs/op
BenchmarkTagMetricHandlerAddsDuplicateTagsLarge-8   	10000000	       121 ns/op	     170 B/op	       1 allocs/op
```

After:
```
goos: darwin
goarch: amd64
pkg: github.com/atlassian/gostatsd/pkg/statsd
BenchmarkTagMetricHandlerAddsDuplicateTagsSmall-8   	 5000000	       407 ns/op	     219 B/op	       3 allocs/op
BenchmarkTagMetricHandlerAddsDuplicateTagsLarge-8   	 2000000	       603 ns/op	     412 B/op	       5 allocs/op
```

2. go test -v -run=Throughput -count=1 ./pkg/statsd

Before:
```
=== RUN   TestStatsdThroughput
--- PASS: TestStatsdThroughput (7.03s)
	statsd_test.go:64: Processed metrics: 280524
			TotalAlloc: 5243524680 (18691 per metric)
			HeapObjects: 1015389
			Mallocs: 93191455 (332 per metric)
			NumGC: 69
			GCCPUFraction: -0.000025
```

After:
```
=== RUN   TestStatsdThroughput
--- PASS: TestStatsdThroughput (7.02s)
	statsd_test.go:64: Processed metrics: 255408
			TotalAlloc: 5871415840 (22988 per metric)
			HeapObjects: 924860
			Mallocs: 112551052 (440 per metric)
			NumGC: 84
			GCCPUFraction: -0.000028
```